api/requirements.txt
fastapi
uvicorn
supabase
google-genai
python-multipart
python-dotenv
slowapi
mercadopago
python-dateutil
